requirements.txt
Django==1.6
PIL==1.1.7
South==0.7.6
Unipath==1.0
bootstrap-admin==0.2.8
distribute==0.7.3
dj-database-url==0.2.2
dj-static==0.0.5
django-tagging==0.3.1
django-tinymce==1.5.2
django-toolbelt==0.0.1
django-wysiwyg-redactor==0.3.8
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
